# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio API
  description: |
    Investment portfolio management API for KFH/BBYN/USA portfolios.
    Supports multi-user isolation, soft-delete, currency conversion (USD→KWD),
    and real-time Yahoo Finance price fetching.

    ## Authentication
    All endpoints (except /health, /api/v1/auth/login, /api/v1/cron/status)
    require a JWT Bearer token obtained via login.

    ## Versioning
    API is versioned via URL prefix `/api/v1/`.

    ## Currencies
    - KFH & BBYN portfolios: KWD
    - USA portfolio: USD
    - All aggregate totals converted to KWD
  version: 1.0.0
  contact:
    name: Sager Alsager
  license:
    name: Private
servers:
  - url: http://localhost:8002
    description: Local development
  - url: https://portfolio-api.onrender.com
    description: Production (Render)
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────
# PATHS
# ──────────────────────────────────────────────
paths:
...
  /api/v1/stocks:
    get:
      operationId: listStocks
      tags: [Securities]
      summary: List user's stocks (with live prices)
      parameters:
        - $ref: "#/components/parameters/PortfolioFilter"
      responses:
        "200":
          description: Stocks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ───── SNAPSHOTS & PERFORMANCE ─────
...
  /api/v1/analysis/stocks:
    get:
      operationId: listAnalysisStocks
      tags: [Peer Analysis]
      summary: List stocks in analysis universe
      responses:
        "200":
          description: Analysis stocks
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnalysisStock"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: addAnalysisStock
      tags: [Peer Analysis]
      summary: Add a stock to analysis universe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisStockCreate"
      responses:
        "201":
          description: Stock added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisStock"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Stock already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/admin/backup/export:
    get:
      operationId: exportBackup
      tags: [Admin]
      summary: Export all user data as Excel
      description: Downloads an Excel workbook with all tables for the current user.
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/admin/backup/import:
    post:
      operationId: importBackup
      tags: [Admin]
      summary: Import data from Excel backup
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                mode:
                  type: string
                  enum: [merge, replace]
                  default: merge
                  description: merge=add missing, replace=wipe and reload
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/v1/auth/login
    CronKeyAuth:
      type: apiKey
      in: header
      name: X-Cron-Key
      description: Server-side cron secret key

  # ───── PARAMETERS ─────
  parameters:
...
    PortfolioFilter:
      name: portfolio
      in: query
      required: false
      schema:
        type: string
        enum: [KFH, BBYN, USA]
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Not found"
...
    Stock:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
          nullable: true
        current_price:
          type: number
          format: double
        portfolio:
          type: string
        currency:
          type: string
        last_updated:
          type: integer
          nullable: true
        price_source:
          type: string
          nullable: true
    StockListResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Stock"

    # --- Snapshots ---
...
    AnalysisStock:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        company_name:
          type: string
        exchange:
          type: string
        currency:
          type: string
        sector:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        outstanding_shares:
          type: number
          format: double
          nullable: true
    AnalysisStockCreate:
      type: object
      required: [symbol, company_name]
      properties:
        symbol:
          type: string
        company_name:
          type: string
        exchange:
          type: string
          default: NYSE
        currency:
          type: string
          default: USD
        sector:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
...
    ImportResult:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            imported:
              type: integer
            skipped:
              type: integer
            errors:
              type: array
              items:
                type: object
                properties:
                  row:
                    type: integer
                  error:
                    type: string
            warnings:
              type: array
              items:
                type: string

    # --- Admin ---
...